+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 `AnimatableStackView` adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.0.0](https://github.com/APUtils/AnimatableStackView/releases/tag/2.0.0)
+Released on 07/09/2019.
+
+#### Changed
+- Create and Configure protocols takes Any as a parameter
+- AnimatableStackView is able to work with any kind of views together
+
+
 ## [1.0.0](https://github.com/APUtils/AnimatableStackView/releases/tag/1.0.0)
 Released on 07/09/2019.